Don't directly diff against upstream. Use merge-base.
diffs should be against the merge-base of upstream and HEAD. Otherwise,
"git diff `git-merge-base HEAD git-svn` git-svn" will get added to the
patch. For example, if a new file has been added since the branch was
rebased, "git-cl upload" will send up a patch which deletes the new file.

I ran into a problem recently where I fetched upstream and didn't bother
rebasing a branch I was working on. (I don't rebase a branch unless it is
old or the new changes are relevant.) When I then tried to 'git-cl upload',
the uploaded patch deleted files which were added since the last rebase of
the branch.

Signed-off-by: Mandeep Singh Baines <msb@google.com>